% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params_get}
\alias{params_get}
\title{Get URL query parameters in Shiny}
\usage{
params_get(session, keys = NULL)
}
\arguments{
\item{session}{Shiny session}

\item{keys}{Optional character vector of keys to pull; if NULL return all}
}
\value{
Named list
}
\description{
Get URL query parameters in Shiny
}
