% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/payload.R
\name{setup_payload_endpoint}
\alias{setup_payload_endpoint}
\title{Setup POST endpoint in server function - MUST be called in server}
\usage{
setup_payload_endpoint(path = "/ingress", session, token = NULL)
}
\arguments{
\item{path}{The URL path to handle POST requests (default "/ingress")}

\item{session}{The Shiny session object}

\item{token}{Optional authentication token for POST requests}
}
\value{
No return value, called for side effects. Registers a POST endpoint
handler with the Shiny session that will process incoming requests.
}
\description{
Setup POST endpoint in server function - MUST be called in server
}
